requirements.txt
snowflake-connector-python>=2.7.0
PyYAML>=6.0
python-dotenv>=0.19.0
spacy>=3.5.0